pyproject.toml
[project]
name = "convert2applevoice"
version = "0.1.0"
description = "Automate Apple Personal Voice creation using TTS output"
authors = [
    { name = "Convert2ApplePVoice Contributors" }
]
dependencies = [
    "pyobjc-framework-Cocoa>=9.2",
    "pyobjc-framework-Vision>=9.2",
    "pytest>=7.4.3",
    "rich>=13.7.0",
    "py3-tts-wrapper[microsoft]>=0.1.0",
    "numpy>=1.24.0",
    "azure-cognitiveservices-speech>=1.31.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"
